version 3.0/requirements.txt
Flask
pymongo
werkzeug
certifi
python-dotenv
